Make amap base request param to base
